- interface fixes - linters issues
