+++ testsCommon/txBuilderStub.go
@@ -1,17 +1,17 @@
 package testsCommon
 
 import (
-	"github.com/ElrondNetwork/elrond-sdk-erdgo/blockchain/cryptoProvider"
+	erdgoCore "github.com/ElrondNetwork/elrond-sdk-erdgo/core"
 	"github.com/ElrondNetwork/elrond-sdk-erdgo/data"
 )
 
 // TxBuilderStub -
 type TxBuilderStub struct {
-	ApplySignatureAndGenerateTxCalled func(cryptoHolder cryptoProvider.CryptoComponentsHolder, arg data.ArgCreateTransaction) (*data.Transaction, error)
+	ApplySignatureAndGenerateTxCalled func(cryptoHolder erdgoCore.CryptoComponentsHolder, arg data.ArgCreateTransaction) (*data.Transaction, error)
 }
 
 // ApplySignatureAndGenerateTx -
-func (stub *TxBuilderStub) ApplySignatureAndGenerateTx(cryptoHolder cryptoProvider.CryptoComponentsHolder, arg data.ArgCreateTransaction) (*data.Transaction, error) {
+func (stub *TxBuilderStub) ApplySignatureAndGenerateTx(cryptoHolder erdgoCore.CryptoComponentsHolder, arg data.ArgCreateTransaction) (*data.Transaction, error) {
 	if stub.ApplySignatureAndGenerateTxCalled != nil {
 		return stub.ApplySignatureAndGenerateTxCalled(cryptoHolder, arg)
 	}
